multi-domain-rag-chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
langchain
langchain-community
langchain-openai
langchain-groq
langchain-huggingface
langchain-text-splitters
sentence-transformers
faiss-cpu
transformers
numpy
pyyaml
streamlit
openai
